+++ packages/uiengine-theme-default/gulpfile.js
@@ -1,25 +1,15 @@
 const gulp = require('gulp')
 const { dirname, join } = require('path')
-const mergeStream = require('merge-stream')
 const runSequence = require('run-sequence')
-const autoprefixer = require('autoprefixer')
-const mqpacker = require('css-mqpacker')
-const csswring = require('csswring')
 const webpack = require('webpack')
 const webpackStream = require('webpack-stream')
 const p = require('gulp-load-plugins')()
 
-const skins = ['default', 'deeplake', 'uiengineering']
 const webpackEnv = process.env.NODE_ENV === 'production' ? 'prod' : 'dev'
 const webpackConfig = require(join(__dirname, `build/webpack.${webpackEnv}.conf.js`))
 
-const paths = {
-  stylesLib: './src/styles/lib'
-}
-
 const src = {
   lib: ['./src/*.js', './src/{__,lib}/**/*.js'], // FIXME: '__' is a hack that allows for lib to be accepted as dynamic path component
-  styles: ['./src/styles/*.styl', './src/styles/components/*.styl'],
   webpack: ['src/{templates,vue}/**/*']
 }
 
@@ -29,24 +19,6 @@ gulp.task('lib', () =>
     .pipe(p.babel())
     .pipe(gulp.dest('./lib'))
 )
-
-const styles = skin =>
-  gulp.src(src.styles)
-    .pipe(p.plumber())
-    .pipe(p.stylus({
-      paths: [paths.stylesLib],
-      import: ['variables', 'mediaQueries', 'mixins', `skins/${skin}`],
-      url: { name: 'embedurl' }
-    }))
-    .pipe(p.concat(`uiengine-${skin}.css`))
-    .pipe(p.postcss([
-      mqpacker,
-      autoprefixer({ browsers: ['last 2 versions'] }),
-      csswring
-    ]))
-    .pipe(gulp.dest('./static/styles'))
-
-gulp.task('styles', () => mergeStream(...skins.map(styles)))
 
 gulp.task('hljs', () =>
   gulp.src(`${dirname(require.resolve('highlight.js/styles/github.css'))}/**`)
@@ -63,8 +35,7 @@ gulp.task('webpack', () =>
 gulp.task('watch', () => {
   gulp.watch(src.lib, ['lib'])
   gulp.watch(src.webpack, ['webpack'])
-  gulp.watch(src.styles.concat([`${paths.stylesLib}/**/*.styl`]), ['styles'])
 })
 
-gulp.task('build', cb => runSequence(['lib', 'styles', 'hljs'], ['webpack'], cb))
+gulp.task('build', cb => runSequence(['lib', 'hljs'], ['webpack'], cb))
 gulp.task('develop', cb => runSequence('build', 'watch', cb))