+++ packages/uiengine-theme-default/src/vue/components/ComponentLabel.vue
@@ -17,3 +17,19 @@ export default {
   }
 }
 </script>
+
+<style lang="stylus" scoped>
+.label
+  padding var(--space-xxs) var(--space-s)
+  border 1px solid var(--color-contentheader-label-bg)
+  border-radius var(--space-m)
+  border-bottom-right-radius 0
+  color var(--color-contentheader-label-text)
+  background var(--color-contentheader-label-bg)
+  font-size var(--font-size-s)
+
+  &--inverted
+    color var(--color-contentheader-label-bg)
+    background var(--color-contentheader-label-text)
+    border-color var(--color-contentheader-label-bg)
+</style>