+++ packages/uiengine-theme-default/src/vue/components/ContentProperties.vue
@@ -49,8 +49,7 @@ export default {
 }
 </script>
 
-<style scoped>
-table {
-  width: 100%;
-}
+<style lang="stylus" scoped>
+table
+  width 100%
 </style>